almost complete implementation, removed custom weight 1 behavior

+++ src/RED4EXT/src/DataStructs/VariantPool.h
@@ -8,7 +8,7 @@ namespace InfiniteRandomizerFramework {
     struct VariantPoolEntry {
         RED4ext::ResourcePath resourcePath;
         std::string appearance;
-        double weight;
+        float weight;
     };
 
     struct VariantPool {
